[Gold III] Title: 감시, Time: 4 ms, Memory: 2220 KB -BaekjoonHub

+++ 백준/Gold/15683. 감시/감시.cc
@@ -81,7 +81,7 @@ int main() {
 		}
 	}
 
-	dfs((int)cams.size(), visited, cams);
+	dfs(cams.size(), visited, cams);
 	cout << ans;
 	return 0;
 }
